requirements.txt
Django==1.9.5
django-bootstrap-dynamic-formsets==0.4.5
django-bootstrap3==7.0.1
django-crispy-forms==1.5.2
django-jet==0.1.4
django-wysiwyg-redactor==0.4.9
gunicorn==19.4.5
Pillow==3.0.0
